refactor(ui): extract message-list layout cache integration
